+++ app/src/main/java/com/lateinit/rightweight/ui/calendar/CalendarFragment.kt
@@ -16,6 +16,7 @@ class CalendarFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         (requireActivity() as HomeActivity).supportActionBar?.setTitle(R.string.calendar)
+        (requireActivity() as HomeActivity).supportActionBar?.setDisplayHomeAsUpEnabled(true)
         return inflater.inflate(R.layout.fragment_calendar, container, false)
     }
