+++ app/src/main/java/com/lateinit/rightweight/ui/home/HomeFragment.kt
@@ -15,6 +15,7 @@ class HomeFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         (requireActivity() as HomeActivity).supportActionBar?.setTitle(R.string.app_name)
+        (requireActivity() as HomeActivity).supportActionBar?.setDisplayHomeAsUpEnabled(true)
         return inflater.inflate(R.layout.fragment_home, container, false)
     }
