+++ app/src/main/java/com/lateinit/rightweight/ui/routine/management/RoutineManagementFragment.kt
@@ -17,6 +17,7 @@ class RoutineManagementFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         (requireActivity() as HomeActivity).supportActionBar?.setTitle(R.string.management)
+        (requireActivity() as HomeActivity).supportActionBar?.setDisplayHomeAsUpEnabled(true)
         return inflater.inflate(R.layout.fragment_routine_management, container, false)
     }
