+++ app/src/main/java/com/lateinit/rightweight/ui/share/detail/SharedRoutineDetailFragment.kt
@@ -3,9 +3,7 @@ package com.lateinit.rightweight.ui.share.detail
 import android.os.Bundle
 import android.view.*
 import android.widget.Button
-import androidx.core.view.MenuProvider
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Lifecycle
 import androidx.navigation.findNavController
 import com.lateinit.rightweight.R
 import com.lateinit.rightweight.ui.home.HomeActivity
@@ -16,17 +14,7 @@ class SharedRoutineDetailFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         (requireActivity() as HomeActivity).supportActionBar?.setTitle(R.string.detail)
-        requireActivity().addMenuProvider(object : MenuProvider {
-            override fun onCreateMenu(menu: Menu, menuInflater: MenuInflater) {
-                // Add menu items here
-                menu.clear()
-                menuInflater.inflate(R.menu.menu_routine_detail, menu)
-            }
-
-            override fun onMenuItemSelected(menuItem: MenuItem): Boolean {
-                return true
-            }
-        }, viewLifecycleOwner, Lifecycle.State.RESUMED)
+        (requireActivity() as HomeActivity).supportActionBar?.setDisplayHomeAsUpEnabled(false)
 
         return inflater.inflate(R.layout.fragment_shared_routine_detail, container, false)
     }
